add back `qlstephen` because its actually useful
it is sometimes necessary to use xattr to get it working properly
though. See the README of qlstephen to figure out how to do this.
Theoretically it would be a good idea to get homebrew to do this itself,
but that might not be possible, and this is in a commit message so I'm
never going to see this again and remember to try to do it.